more auth recipe abstraction
